requirements.txt
bs4
uvicorn
fastapi
requests
beautifulsoup4
python-telegram-bot